Move axios into a single file
