+++ src/redux/actions/actionHelper.js
@@ -1,6 +1,6 @@
 import Axios from 'axios';
 import baseURL from './baseURL';
-
+/* eslint-disable camelcase */
 const queryBuilder = () => {
   const get = (address, type) => (dispatch) => {
     const token = localStorage.getItem('jwt');
@@ -141,8 +141,125 @@ const queryBuilder = () => {
         payload: err,
       }));
   };
+
+  const postFavourites = (address, data, user) => (dispatch) => {
+    const token = localStorage.getItem('jwt');
+    const authAxios = Axios.create({
+      baseURL: `${baseURL}`,
+      headers: {
+        Authorization: `Bearer ${token}`,
+        'Content-Type': 'application/json',
+      },
+    });
+
+    const cabinData = {
+      favourite: data,
+    };
+    const payload = {
+      message: 'Cabin was successfully added to Favourites!',
+      type: 'add_fav',
+    };
+
+    authAxios
+      .post(address, cabinData)
+      .then((res) => {
+        dispatch({
+          type: 'ADD_FAVOURITE',
+          payload: res.data.cabin_id,
+        });
+
+        dispatch({
+          type: 'FETCH_CABIN',
+          payload: res.data.cabin_id,
+        });
+
+        dispatch({
+          type: 'FAVOURITE_CABIN',
+        });
+
+        dispatch({
+          type: 'SUCCESS_MESSAGE',
+          payload,
+        });
+
+        dispatch({
+          type: 'FETCH_USER',
+          payload: user,
+        });
+      })
+      .catch((err) => dispatch({
+        type: 'CREATE_ERROR',
+        payload: err,
+      }));
+  };
+
+  const removeFavourites = (address, cabin_id, user) => (dispatch) => {
+    const token = localStorage.getItem('jwt');
+    const authAxios = Axios.create({
+      baseURL: `${baseURL}`,
+      headers: {
+        Authorization: `Bearer ${token}`,
+        'Content-Type': 'application/json',
+      },
+    });
+
+    const payload = {
+      message: 'Cabin was successfully removed from Favourites!',
+      type: 'remove_fav',
+    };
+    authAxios
+      .delete(address)
+      .then(() => {
+        dispatch({
+          type: 'REMOVE_FAVOURITE',
+          payload: cabin_id,
+        });
+
+        dispatch({
+          type: 'NOT_FAVOURITE_CABIN',
+        });
+
+        dispatch({
+          type: 'SUCCESS_MESSAGE',
+          payload,
+        });
+
+        dispatch({
+          type: 'FETCH_USER',
+          payload: user,
+        });
+      })
+      .catch((err) => dispatch({
+        type: 'CREATE_ERROR',
+        payload: err,
+      }));
+  };
+
+  const postImage = (image) => (dispatch) => {
+    const token = localStorage.getItem('jwt');
+    const userAxios = Axios.create({
+      baseURL: `${baseURL}`,
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    });
+
+    userAxios
+      .post('/api/v1/image_uploaders.json', image)
+      .then((res) => {
+        dispatch({
+          type: 'ADD_IMAGE',
+          payload: res.data,
+        });
+      })
+      .catch((err) => dispatch({
+        type: 'CREATE_ERROR',
+        payload: err,
+      }));
+  };
+
   return {
-    get, post, update, deletes, postUser,
+    get, post, update, deletes, postUser, postFavourites, removeFavourites, postImage,
   };
 };
